YakSH bytecode VM
I need some way to call built-in functions. I'll probably end up with a
syscall opcode or something similar. Ignoring how rigid the language
already is, I'm beginning to see the tradeoffs between ease of
implementing the compiler and ease of programming in a language. It's not
a direct connection (some things that make it easy to program are also
easy to implement), but I recognize it's easy to take shortcuts in the
compiler and expect the programmer to make up the difference in their
code.